Organize files and improve SendEmailToAdmin.php
