+++ src/plugins/LifePerformancesPlugin/includes/DeleteVideo.php
